Add canOverrideExistingModule() to support RN 0.60+ (#7)
This fixes an issue where React Native refuses to load citing

```
Native module DevSettings tried to override com.facebook.react.modules.debug.DevSettingsModule. Check that getPackages() method in MainApplication.java, it might be that module is being created twice. If this was your intention, set canOverrideExistingModule=true
```

It is the intention!